Remove old oauth services
