+++ services/auth/service/github_service.go
@@ -2,44 +2,42 @@ package service
 
 import (
 	"errors"
+	"strconv"
+	"strings"
 	"github.com/HackIllinois/api/services/auth/config"
 	"github.com/HackIllinois/api/services/auth/models"
 	"github.com/levigross/grequests"
-	"strconv"
 )
 
-/*
-	Uses a valid oauth token to get the user's primary email
-*/
-func GetGithubEmail(oauth_token string) (string, bool, error) {
-	request, err := grequests.Get("https://api.github.com/user/emails", &grequests.RequestOptions{
-		Headers: map[string]string{"Authorization": "token " + oauth_token},
-	})
+type GitHubOAuthProvider struct {
+	token string
+	isVerifiedUser bool
+}
 
-	if err != nil {
-		return "", false, err
+func NewGitHubOAuth() *GitHubOAuthProvider {
+	return &GitHubOAuthProvider{
+		token: "",
+		isVerifiedUser: false,
 	}
-
-	var emails []models.GithubEmail
-	err = request.JSON(&emails)
-
-	if err != nil {
-		return "", false, err
-	}
-
-	for _, email := range emails {
-		if email.IsPrimary {
-			return email.Email, email.IsVerified, nil
-		}
-	}
-
-	return "", false, errors.New("No primary email")
 }
 
 /*
-	Uses a valid oauth code to get a valid oauth token for the user
+	Returns the url to redirects to for OAuth authorization
 */
-func GetGithubOauthToken(code string) (string, error) {
+func (provider *GitHubOAuthProvider) GetAuthorizationRedirect(redirect_uri string) (string, error) {
+	return ConstructSafeURL("https", "github.com", "login/oauth/authorize",
+		map[string]string{
+			"client_id":    config.GITHUB_CLIENT_ID,
+			"scope":        "user:email",
+			"redirect_uri": redirect_uri,
+		},
+	)
+}
+
+/*
+	Exchanges an OAuth code for an OAuth token
+*/
+func (provider *GitHubOAuthProvider) Authorize(code string, redirect_uri string) error {
 	request, err := grequests.Post("https://github.com/login/oauth/access_token", &grequests.RequestOptions{
 		Params: map[string]string{
 			"client_id":     config.GITHUB_CLIENT_ID,
@@ -52,97 +50,95 @@ func GetGithubOauthToken(code string) (string, error) {
 	})
 
 	if err != nil {
-		return "", err
+		return err
 	}
 
 	var oauth_token models.GithubOauthToken
 	err = request.JSON(&oauth_token)
 
 	if err != nil {
-		return "", err
+		return err
 	}
 
 	if oauth_token.Token == "" {
-		return "", errors.New("Invalid oauth code")
+		return errors.New("Invalid oauth code")
 	}
 
-	return oauth_token.Token, nil
+	provider.token = oauth_token.Token
+
+	return nil
 }
 
 /*
-	Uses a valid oauth token to get the user's unique id
+	Retrieves user info from the OAuth provider
 */
-func GetGithubUniqueId(oauth_token string) (string, error) {
+func (provider *GitHubOAuthProvider) GetUserInfo() (*models.UserInfo, error) {
+	var user_info models.UserInfo
+
 	request, err := grequests.Get("https://api.github.com/user", &grequests.RequestOptions{
-		Headers: map[string]string{"Authorization": "token " + oauth_token},
+		Headers: map[string]string{
+			"Authorization": "token " + provider.token,
+		},
 	})
 
 	if err != nil {
-		return "", err
+		return nil, err
 	}
 
-	var user_info models.GithubUserInfo
-	err = request.JSON(&user_info)
+	var github_user_info models.GithubUserInfo
+	err = request.JSON(&github_user_info)
 
 	if err != nil {
-		return "", err
+		return nil, err
 	}
 
-	if user_info.ID == 0 {
-		return "", errors.New("Invalid oauth token")
+	if github_user_info.ID == 0 {
+		return nil, errors.New("Invalid oauth token")
 	}
 
-	return "github" + strconv.Itoa(user_info.ID), nil
+	user_info.Username = github_user_info.Username
+	user_info.ID = "github" + strconv.Itoa(github_user_info.ID)
+
+	split_name := strings.SplitAfterN(github_user_info.Name, " ", 2)
+
+	user_info.FirstName = strings.TrimSpace(split_name[0])
+
+	if len(split_name) < 2 {
+		user_info.LastName = ""
+	} else {
+		user_info.LastName = strings.TrimSpace(split_name[1])
+	}
+
+	request, err = grequests.Get("https://api.github.com/user/emails", &grequests.RequestOptions{
+		Headers: map[string]string{
+			"Authorization": "token " + provider.token,
+		},
+	})
+
+	if err != nil {
+		return nil, err
+	}
+
+	var emails []models.GithubEmail
+	err = request.JSON(&emails)
+
+	if err != nil {
+		return nil, err
+	}
+
+	for _, email := range emails {
+		if email.IsPrimary {
+			user_info.Email = email.Email
+			provider.isVerifiedUser = email.IsVerified
+		}
+	}
+
+	return &user_info, nil
 }
 
 /*
-	Uses a valid oauth token to get the user's username
+	Returns true if the user has a verified email
 */
-func GetGithubUsername(oauth_token string) (string, error) {
-	request, err := grequests.Get("https://api.github.com/user", &grequests.RequestOptions{
-		Headers: map[string]string{"Authorization": "token " + oauth_token},
-	})
-
-	if err != nil {
-		return "", err
-	}
-
-	var user_info models.GithubUserInfo
-	err = request.JSON(&user_info)
-
-	if err != nil {
-		return "", err
-	}
-
-	if user_info.ID == 0 {
-		return "", errors.New("Invalid oauth token")
-	}
-
-	return user_info.Username, nil
-}
-
-/*
-	Uses a valid oauth token to get the user's name
-*/
-func GetGithubName(oauth_token string) (string, error) {
-	request, err := grequests.Get("https://api.github.com/user", &grequests.RequestOptions{
-		Headers: map[string]string{"Authorization": "token " + oauth_token},
-	})
-
-	if err != nil {
-		return "", err
-	}
-
-	var user_info models.GithubUserInfo
-	err = request.JSON(&user_info)
-
-	if err != nil {
-		return "", err
-	}
-
-	if user_info.ID == 0 {
-		return "", errors.New("Invalid oauth token")
-	}
-
-	return user_info.Name, nil
+func (provider *GitHubOAuthProvider) IsVerifiedUser() bool {
+	return provider.isVerifiedUser
 }
